+++ Reddit/Easy/alphabet_cipher/src/lib.rs
@@ -13,12 +13,10 @@ pub fn decrypt(code: &str, keyword: &str) -> String {
     let code: Vec<char> = code.chars().collect();
     let keyword: Vec<char> = keyword.chars().collect();
 
-    let mut result = String::new();
-    for i in 0..code.len() {
-        result.push((97 + (26 + code[i] as i32 - keyword[i%keyword.len()] as i32)as u8 %26u8) as char);
-    }
-
-    result
+    (0..code.len())
+        .map(|i| (97 + (26 + code[i] as i32 -
+                    keyword[i%keyword.len()] as i32)as u8 %26u8) as char)
+        .collect()
 }
 
 #[cfg(test)]